yolov5/requirements.txt
blinker==1.7.0
boto3==1.34.80
botocore==1.34.80
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
command-not-found==0.3
contourpy==1.2.1
cryptography==3.4.8
cycler==0.12.1
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.1
filelock==3.13.3
Flask==3.0.3
fonttools==4.51.0
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
httplib2==0.20.2
idna==3.6
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.3
jmespath==1.0.1
keyring==23.5.0
kiwisolver==1.4.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.1.5
matplotlib==3.8.4
more-itertools==8.10.0
mpmath==1.3.0
netifaces==0.11.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.1
pillow==10.3.0
psutil==5.9.8
py-cpuinfo==9.0.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==5.4.1
requests==2.31.0
s3transfer==0.10.1
scipy==1.13.0
seaborn==0.13.2
SecretStorage==3.3.1
six==1.16.0
smmap==5.0.1
sympy==1.12
systemd-python==234
thop==0.1.1.post2209072238
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
triton==2.2.0
typing_extensions==4.11.0
tzdata==2024.1
ubuntu-advantage-tools==8001
ufw==0.36.1
ultralytics==8.1.45
unattended-upgrades==0.1
urllib3==2.2.1
wadllib==1.3.6
Werkzeug==3.0.2
zipp==1.0.0
